requirements.txt
flask
torch
torchvision
pillow
transformers
flask_cors
